+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- BREAKING CHANGE: change `spec.containers.*.env.*.values_from` -> `spec.containers.*.env.*.value_from`
+- BREAKING CHANGE: change `spec.containers.*.ports.*.protocol` defaulting from "TCP" to `null`
+
 ## [0.1.0] BREAKING CHANGES
 
 ### Added